+++ zetaclient/chains/evm/signer/signer.go
@@ -362,6 +362,12 @@ func (signer *Signer) TryProcessOutbound(
 	zetacoreClient interfaces.ZetacoreClient,
 	height uint64,
 ) {
+	app, err := zctx.FromContext(ctx)
+	if err != nil {
+		signer.Logger().Std.Error().Err(err).Msg("error getting app context")
+		return
+	}
+
 	// end outbound process on panic
 	defer func() {
 		outboundProc.EndTryProcess(outboundID)
@@ -396,6 +402,17 @@ func (signer *Signer) TryProcessOutbound(
 		return
 	}
 	if skipTx {
+		return
+	}
+
+	toChain, err := app.GetChain(txData.toChainID.Int64())
+	switch {
+	case err != nil:
+		logger.Error().Err(err).Msgf("error getting toChain %d", txData.toChainID.Int64())
+		return
+	case toChain.IsZeta():
+		// should not happen
+		logger.Error().Msgf("unable to TryProcessOutbound when toChain is zetaChain (%d)", toChain.ID())
 		return
 	}
 
@@ -447,7 +464,7 @@ func (signer *Signer) TryProcessOutbound(
 			logger.Info().Msgf(
 				"SignWithdrawTx: %d => %d, nonce %d, gasPrice %d",
 				cctx.InboundParams.SenderChainId,
-				txData.toChainID.Int64(),
+				toChain.ID(),
 				cctx.GetCurrentOutboundParam().TssNonce,
 				txData.gasPrice,
 			)
@@ -456,7 +473,7 @@ func (signer *Signer) TryProcessOutbound(
 			logger.Info().Msgf(
 				"SignERC20WithdrawTx: %d => %d, nonce %d, gasPrice %d",
 				cctx.InboundParams.SenderChainId,
-				txData.toChainID.Int64(),
+				toChain.ID(),
 				cctx.GetCurrentOutboundParam().TssNonce,
 				txData.gasPrice,
 			)
@@ -465,7 +482,7 @@ func (signer *Signer) TryProcessOutbound(
 			logger.Info().Msgf(
 				"SignOutbound: %d => %d, nonce %d, gasPrice %d",
 				cctx.InboundParams.SenderChainId,
-				txData.toChainID.Int64(),
+				toChain.ID(),
 				cctx.GetCurrentOutboundParam().TssNonce,
 				txData.gasPrice,
 			)
@@ -481,7 +498,7 @@ func (signer *Signer) TryProcessOutbound(
 			logger.Info().Msgf(
 				"SignRevertTx: %d => %d, nonce %d, gasPrice %d",
 				cctx.InboundParams.SenderChainId,
-				txData.toChainID.Int64(), cctx.GetCurrentOutboundParam().TssNonce,
+				toChain.ID(), cctx.GetCurrentOutboundParam().TssNonce,
 				txData.gasPrice,
 			)
 			txData.srcChainID = big.NewInt(cctx.OutboundParams[0].ReceiverChainId)
@@ -491,7 +508,7 @@ func (signer *Signer) TryProcessOutbound(
 			logger.Info().Msgf(
 				"SignWithdrawTx: %d => %d, nonce %d, gasPrice %d",
 				cctx.InboundParams.SenderChainId,
-				txData.toChainID.Int64(),
+				toChain.ID(),
 				cctx.GetCurrentOutboundParam().TssNonce,
 				txData.gasPrice,
 			)
@@ -499,7 +516,7 @@ func (signer *Signer) TryProcessOutbound(
 		case coin.CoinType_ERC20:
 			logger.Info().Msgf("SignERC20WithdrawTx: %d => %d, nonce %d, gasPrice %d",
 				cctx.InboundParams.SenderChainId,
-				txData.toChainID.Int64(),
+				toChain.ID(),
 				cctx.GetCurrentOutboundParam().TssNonce,
 				txData.gasPrice,
 			)
@@ -513,7 +530,7 @@ func (signer *Signer) TryProcessOutbound(
 		logger.Info().Msgf(
 			"SignRevertTx: %d => %d, nonce %d, gasPrice %d",
 			cctx.InboundParams.SenderChainId,
-			txData.toChainID.Int64(),
+			toChain.ID(),
 			cctx.GetCurrentOutboundParam().TssNonce,
 			txData.gasPrice,
 		)
@@ -529,7 +546,7 @@ func (signer *Signer) TryProcessOutbound(
 		logger.Info().Msgf(
 			"SignOutbound: %d => %d, nonce %d, gasPrice %d",
 			cctx.InboundParams.SenderChainId,
-			txData.toChainID.Int64(),
+			toChain.ID(),
 			cctx.GetCurrentOutboundParam().TssNonce,
 			txData.gasPrice,
 		)
@@ -543,7 +560,7 @@ func (signer *Signer) TryProcessOutbound(
 	logger.Info().Msgf(
 		"Key-sign success: %d => %d, nonce %d",
 		cctx.InboundParams.SenderChainId,
-		txData.toChainID.Int64(),
+		toChain.ID(),
 		cctx.GetCurrentOutboundParam().TssNonce,
 	)
 
@@ -567,14 +584,16 @@ func (signer *Signer) BroadcastOutbound(
 		return
 	}
 
-	// Get destination chain for logging
-	toChain, found := chains.GetChainFromChainID(txData.toChainID.Int64(), app.GetAdditionalChains())
-	if !found {
-		logger.Warn().Msgf("BroadcastOutbound: unknown chain %d", txData.toChainID.Int64())
+	toChain, err := app.GetChain(txData.toChainID.Int64())
+	switch {
+	case err != nil:
+		logger.Error().Err(err).Msgf("error getting toChain %d", txData.toChainID.Int64())
 		return
-	}
-
-	if tx == nil {
+	case toChain.IsZeta():
+		// should not happen
+		logger.Error().Msgf("unable to broadcast when toChain is zetaChain (%d)", toChain.ID())
+		return
+	case tx == nil:
 		logger.Warn().Msgf("BroadcastOutbound: no tx to broadcast %s", cctx.Index)
 		return
 	}
@@ -591,15 +610,15 @@ func (signer *Signer) BroadcastOutbound(
 			log.Warn().
 				Err(err).
 				Msgf("BroadcastOutbound: error broadcasting tx %s on chain %d nonce %d retry %d signer %s",
-					outboundHash, toChain.ChainId, cctx.GetCurrentOutboundParam().TssNonce, i, myID)
+					outboundHash, toChain.ID(), cctx.GetCurrentOutboundParam().TssNonce, i, myID)
 			retry, report := zetacore.HandleBroadcastError(
 				err,
 				strconv.FormatUint(cctx.GetCurrentOutboundParam().TssNonce, 10),
-				toChain.String(),
+				fmt.Sprintf("%d", toChain.ID()),
 				outboundHash,
 			)
 			if report {
-				signer.reportToOutboundTracker(ctx, zetacoreClient, toChain.ChainId, tx.Nonce(), outboundHash, logger)
+				signer.reportToOutboundTracker(ctx, zetacoreClient, toChain.ID(), tx.Nonce(), outboundHash, logger)
 			}
 			if !retry {
 				break
@@ -608,8 +627,8 @@ func (signer *Signer) BroadcastOutbound(
 			continue
 		}
 		logger.Info().Msgf("BroadcastOutbound: broadcasted tx %s on chain %d nonce %d signer %s",
-			outboundHash, toChain.ChainId, cctx.GetCurrentOutboundParam().TssNonce, myID)
-		signer.reportToOutboundTracker(ctx, zetacoreClient, toChain.ChainId, tx.Nonce(), outboundHash, logger)
+			outboundHash, toChain.ID(), cctx.GetCurrentOutboundParam().TssNonce, myID)
+		signer.reportToOutboundTracker(ctx, zetacoreClient, toChain.ID(), tx.Nonce(), outboundHash, logger)
 		break // successful broadcast; no need to retry
 	}
 }